style: 중복되는 컴포넌트 분리
